requirements.txt
annotated-types==0.6.0
asgiref==3.7.2
attrs==23.1.0
beautifulsoup4==4.12.2
branca==0.6.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
class-registry==2.1.2
distlib==0.3.7
Django==4.2.6
django-bootstrap-datepicker-plus==5.0.4
django-filter==23.3
django-js-asset==2.1.0
django-mathfilters==1.0.0
django-mptt==0.15.0
django-spurl==0.6.8
filelock==3.12.4
filters==1.3.2
folium==0.14.0
geographiclib==2.0
geopy==2.4.0
h11==0.14.0
idna==3.4
Jinja2==3.1.2
leaflet==0.0.3
lxml==4.9.3
MarkupSafe==2.1.3
numpy==1.26.1
outcome==1.3.0.post0
pathlib==1.0.1
Pillow==10.1.0
platformdirs==3.11.0
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
PySocks==1.7.1
python-dateutil==2.8.2
python-docx==1.0.1
pytz==2023.3.post1
regex==2023.10.3
requests==2.31.0
selenium==4.15.2
six==1.16.0
sniffio==1.3.0
sorl-thumbnail==12.10.0
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
trio==0.23.1
trio-websocket==0.11.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
URLObject==2.4.3
virtualenv==20.24.6
virtualenvwrapper-win==1.2.7
wsproto==1.2.0
